03-data-lake-with-google-cloud-storage/examples/uploading-to-gcs/pyproject.toml
[tool.poetry]
name = "uploading-to-gcs"
version = "0.1.0"
description = ""
authors = ["zkan <zkan.cs@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
google-cloud-storage = "^2.8.0"
scrapy = "^2.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
